refactor: refacot of folder and names mangadex
